Refactor: move DTTP related jobs into dttp folder
