--- app/jobs/change_trainee_status_job.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-class ChangeTraineeStatusJob < ApplicationJob
-  queue_as :default
-  retry_on Dttp::Client::HttpError
-
-  def perform(trainee, status, entity_type)
-    Dttp::UpdateTraineeStatus.call(status: status, trainee: trainee, entity_type: entity_type)
-  end
-end
