+++ app/jobs/dttp/recommend_for_award_job.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+module Dttp
+  class RecommendForAwardJob < ApplicationJob
+    queue_as :default
+    retry_on Client::HttpError
+
+    def perform(trainee)
+      RecommendForAward.call(trainee: trainee)
+
+      UpdateTraineeStatus.call(
+        status: DttpStatuses::STANDARDS_MET,
+        trainee: trainee,
+        entity_type: UpdateTraineeStatus::CONTACT_ENTITY_TYPE,
+      )
+
+      Dttp::UpdateTraineeStatus.call(
+        status: DttpStatuses::STANDARDS_MET,
+        trainee: trainee,
+        entity_type: UpdateTraineeStatus::PLACEMENT_ASSIGNMENT_ENTITY_TYPE,
+      )
+    end
+  end
+end
