+++ app/jobs/dttp/register_for_trn_job.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+module Dttp
+  class RegisterForTrnJob < ApplicationJob
+    queue_as :default
+    retry_on Client::HttpError
+
+    def perform(trainee, created_by_dttp_id)
+      RegisterForTrn.call(trainee: trainee, created_by_dttp_id: created_by_dttp_id)
+
+      ChangeTraineeStatusJob.perform_later(
+        trainee,
+        DttpStatuses::PROSPECTIVE_TRAINEE_TRN_REQUESTED,
+        UpdateTraineeStatus::CONTACT_ENTITY_TYPE,
+      )
+
+      ChangeTraineeStatusJob.perform_later(
+        trainee,
+        DttpStatuses::PROSPECTIVE_TRAINEE_TRN_REQUESTED,
+        UpdateTraineeStatus::PLACEMENT_ASSIGNMENT_ENTITY_TYPE,
+      )
+    end
+  end
+end
