+++ app/jobs/dttp/sync_providers_job.rb
@@ -7,9 +7,9 @@ module Dttp
     def perform(request_uri = nil)
       return unless FeatureService.enabled?(:sync_from_dttp)
 
-      @provider_list = Dttp::RetrieveProviders.call(request_uri: request_uri)
+      @provider_list = RetrieveProviders.call(request_uri: request_uri)
 
-      Dttp::Provider.upsert_all(provider_attributes, unique_by: :dttp_id)
+      Provider.upsert_all(provider_attributes, unique_by: :dttp_id)
 
       Dttp::SyncProvidersJob.perform_later(next_page_url) if has_next_page?
     end
@@ -19,7 +19,7 @@ module Dttp
     attr_reader :provider_list
 
     def provider_attributes
-      Dttp::Parsers::Account.to_provider_attributes(providers: provider_list[:items])
+      Parsers::Account.to_provider_attributes(providers: provider_list[:items])
     end
 
     def next_page_url
