+++ app/jobs/dttp/sync_users_job.rb
@@ -7,11 +7,11 @@ module Dttp
     def perform(request_uri = nil)
       return unless FeatureService.enabled?(:sync_from_dttp)
 
-      @user_list = Dttp::RetrieveUsers.call(request_uri: request_uri)
+      @user_list = RetrieveUsers.call(request_uri: request_uri)
 
-      Dttp::User.upsert_all(user_attributes, unique_by: :dttp_id)
+      User.upsert_all(user_attributes, unique_by: :dttp_id)
 
-      Dttp::SyncUsersJob.perform_later(next_page_url) if next_page_url.present?
+      SyncUsersJob.perform_later(next_page_url) if next_page_url.present?
     end
 
   private
@@ -19,7 +19,7 @@ module Dttp
     attr_reader :user_list
 
     def user_attributes
-      Dttp::Parsers::Contact.to_user_attributes(contacts: user_list[:items])
+      Parsers::Contact.to_user_attributes(contacts: user_list[:items])
     end
 
     def next_page_url
