+++ app/jobs/dttp/update_trainee_to_dttp_job.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Dttp
+  class UpdateTraineeToDttpJob < ApplicationJob
+    queue_as :dttp
+
+    def perform(trainee)
+      ContactUpdate.call(trainee: trainee)
+    end
+  end
+end
