--- app/jobs/update_trainee_to_dttp_job.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class UpdateTraineeToDttpJob < ApplicationJob
-  queue_as :dttp
-
-  def perform(trainee)
-    Dttp::ContactUpdate.call(trainee: trainee)
-  end
-end
