+++ spec/controllers/trainees/award_recommendations_controller_spec.rb
@@ -17,11 +17,11 @@ describe Trainees::AwardRecommendationsController do
     it "updates the placement assignment in DTTP to mark it ready for QTS" do
       expect {
         post :create, params: { trainee_id: trainee }
-      }.to have_enqueued_job(RecommendForAwardJob).with(trainee)
+      }.to have_enqueued_job(Dttp::RecommendForAwardJob).with(trainee)
     end
 
     it "queues a background job to poll for the trainee's QTS" do
-      expect(RetrieveAwardJob).to receive(:perform_with_default_delay).with(trainee)
+      expect(Dttp::RetrieveAwardJob).to receive(:perform_with_default_delay).with(trainee)
       post :create, params: { trainee_id: trainee }
     end
