+++ spec/controllers/trainees/confirm_withdrawals_controller_spec.rb
@@ -17,7 +17,7 @@ describe Trainees::ConfirmWithdrawalsController do
     it "updates the placement assignment in DTTP to mark it as withdrawn" do
       expect {
         post :update, params: { trainee_id: trainee }
-      }.to have_enqueued_job(WithdrawJob).with(trainee)
+      }.to have_enqueued_job(Dttp::WithdrawJob).with(trainee)
     end
 
     context "trainee state" do
