+++ spec/jobs/dttp/sync_users_job_spec.rb
@@ -21,7 +21,7 @@ module Dttp
 
     before do
       enable_features(:sync_from_dttp)
-      allow(Dttp::RetrieveUsers).to receive(:call) { user_list }
+      allow(RetrieveUsers).to receive(:call) { user_list }
     end
 
     it "enqueues job with the next_page_url" do
