+++ spec/support/shared_examples/run_create_or_update_consistency_check_job.rb
@@ -7,7 +7,7 @@ shared_examples "CreateOrUpdateConsistencyCheckJob" do |model|
     it "enqueues the CreateOrUpdateConsistencyJob" do
       expect {
         model.call(trainee: trainee)
-      }.to have_enqueued_job(CreateOrUpdateConsistencyCheckJob).with(trainee)
+      }.to have_enqueued_job(Dttp::CreateOrUpdateConsistencyCheckJob).with(trainee)
     end
   end
 end
